rename costing/units directory to costing/unit_models
