+++ watertap/costing/unit_models/energy_recovery_device.py
