+++ watertap/unit_models/anaerobic_digestor.py
@@ -67,7 +67,7 @@ from idaes.core.util.constants import Constants
 from idaes.core.util.exceptions import ConfigurationError, InitializationError
 from idaes.core.util.tables import create_stream_table_dataframe
 
-from watertap.costing.units.anaerobic_digestor import cost_anaerobic_digestor
+from watertap.costing.unit_models.anaerobic_digestor import cost_anaerobic_digestor
 
 __author__ = "Alejandro Garciadiego, Andrew Lee, Xinhong Liu"
