+++ watertap/unit_models/crystallizer.py
@@ -40,7 +40,7 @@ import idaes.core.util.scaling as iscale
 import idaes.logger as idaeslog
 
 from watertap.core import InitializationMixin
-from watertap.costing.units.crystallizer import cost_crystallizer
+from watertap.costing.unit_models.crystallizer import cost_crystallizer
 
 _log = idaeslog.getLogger(__name__)
