+++ watertap/unit_models/electrodialysis_1D.py
@@ -50,7 +50,7 @@ import idaes.logger as idaeslog
 from enum import Enum
 
 from watertap.core import ControlVolume1DBlock, InitializationMixin
-from watertap.costing.units.electrodialysis import cost_electrodialysis
+from watertap.costing.unit_models.electrodialysis import cost_electrodialysis
 
 __author__ = "Xiangyu Bi, Austin Ladshaw"
