+++ watertap/unit_models/mvc/components/compressor.py
@@ -36,7 +36,7 @@ import idaes.core.util.scaling as iscale
 import idaes.logger as idaeslog
 
 from watertap.core import InitializationMixin
-from watertap.costing.units.compressor import cost_compressor
+from watertap.costing.unit_models.compressor import cost_compressor
 
 
 _log = idaeslog.getLogger(__name__)
