+++ watertap/unit_models/nanofiltration_0D.py
@@ -41,7 +41,7 @@ import idaes.core.util.scaling as iscale
 import idaes.logger as idaeslog
 
 from watertap.core import InitializationMixin
-from watertap.costing.units.nanofiltration import cost_nanofiltration
+from watertap.costing.unit_models.nanofiltration import cost_nanofiltration
 
 
 _log = idaeslog.getLogger(__name__)
