+++ watertap/unit_models/nanofiltration_ZO.py
@@ -40,7 +40,7 @@ import idaes.core.util.scaling as iscale
 import idaes.logger as idaeslog
 
 from watertap.core import ControlVolume0DBlock, InitializationMixin
-from watertap.costing.units.nanofiltration import cost_nanofiltration
+from watertap.costing.unit_models.nanofiltration import cost_nanofiltration
 
 
 _log = idaeslog.getLogger(__name__)
