+++ watertap/unit_models/pressure_exchanger.py
@@ -38,7 +38,7 @@ from idaes.core.util.tables import create_stream_table_dataframe
 import idaes.core.util.scaling as iscale
 
 from watertap.core import ControlVolume0DBlock, InitializationMixin
-from watertap.costing.units.pressure_exchanger import cost_pressure_exchanger
+from watertap.costing.unit_models.pressure_exchanger import cost_pressure_exchanger
 
 _log = idaeslog.getLogger(__name__)
